feat(ui): implement filter ux
- enables multiple exact userProperty matches
- adds visual ux
